requirements.txt
langgraph>=0.2.50
langchain-core>=0.3.0
google-genai>=1.0.0
pydantic>=2.6.0
httpx>=0.27.0
python-dotenv>=1.0.0
boto3>=1.34.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
